+++ app/services/redirects/IncomeFromStatePensionsPages.scala
@@ -16,37 +16,37 @@
 
 package services.redirects
 
-sealed trait IncomeFromPensionsPages {
+sealed trait IncomeFromStatePensionsPages {
   val journeyNo: Int
 }
 
-object IncomeFromPensionsPages {
+object IncomeFromStatePensionsPages {
 
-  case object DoYouGetRegularStatePaymentsPage extends IncomeFromPensionsPages {
+  case object DoYouGetRegularStatePaymentsPage extends IncomeFromStatePensionsPages {
     override val journeyNo: Int = 1
   }
 
-  case object WhenDidYouStartGettingStatePaymentsPage extends IncomeFromPensionsPages {
+  case object WhenDidYouStartGettingStatePaymentsPage extends IncomeFromStatePensionsPages {
     override val journeyNo: Int = 2
   }
 
-  case object StatePensionLumpSumPage extends IncomeFromPensionsPages {
+  case object StatePensionLumpSumPage extends IncomeFromStatePensionsPages {
     override val journeyNo: Int = 3
   }
 
-  case object TaxOnStatePensionLumpSumPage extends IncomeFromPensionsPages {
+  case object TaxOnStatePensionLumpSumPage extends IncomeFromStatePensionsPages {
     override val journeyNo: Int = 4
   }
 
-  case object WhenDidYouGetYourStatePensionLumpSumPage extends IncomeFromPensionsPages {
+  case object WhenDidYouGetYourStatePensionLumpSumPage extends IncomeFromStatePensionsPages {
     override val journeyNo: Int = 5
   }
 
-  case object AddStatePensionToIncomeTaxCalcPage extends IncomeFromPensionsPages {
+  case object AddStatePensionToIncomeTaxCalcPage$State extends IncomeFromStatePensionsPages {
     override val journeyNo: Int = 6
   }
 
-  case object StatePensionsCYAPage extends IncomeFromPensionsPages {
+  case object StateStatePensionsCYAPage$ extends IncomeFromStatePensionsPages {
     override val journeyNo: Int = 7
   }
 }
